requirements.txt
# requirements.txt

Flask
pandas
Pillow
pytesseract
flask-cors